+++ src/components/Forms/RecruitForm/Groups/RecruitBasicInfo/index.tsx
@@ -1,13 +1,9 @@
-import type { CSSProperties, ReactNode } from 'react';
-import type { IconNames } from '~/components/Common';
 import type { RecruitFormValues } from '~/components/Forms/RecruitForm/utils';
 
-import { css } from '@emotion/react';
 import { useWatch } from 'react-hook-form';
 
-import { Accordion, Icon } from '~/components/Common';
 import FieldOverview from '~/components/Forms/RecruitForm/Common/FieldOverview';
-import { recruitFormExpandCss } from '~/components/Forms/RecruitForm/Common/recruitFormExpandCss';
+import { RecruitFormAccordion } from '~/components/Forms/RecruitForm/Common/RecruitFormAccordion';
 import SelectedSkills from '~/components/Forms/RecruitForm/Common/SelectedSkills';
 import {
   EndDate,
@@ -15,7 +11,6 @@ import {
   Skills,
 } from '~/components/Forms/RecruitForm/Fields';
 import { RecruitCategoryName } from '~/components/Forms/RecruitForm/utils';
-import { flex, fontCss, palettes } from '~/styles/utils';
 
 interface RecruitBasicInfoProps {
   className?: string;
@@ -26,13 +21,13 @@ export const RecruitBasicInfo = (props: RecruitBasicInfoProps) => {
     <div {...props}>
       <FieldOverview>리쿠르팅 기본 정보</FieldOverview>
 
-      <Accordion.Root type="multiple" css={accordionRootCss}>
+      <RecruitFormAccordion.Root>
         <ParticipantsLayer />
 
         <EndDateLayer />
 
         <SkillsLayer />
-      </Accordion.Root>
+      </RecruitFormAccordion.Root>
     </div>
   );
 };
@@ -58,14 +53,14 @@ const ParticipantsLayer = () => {
     .join(', ');
 
   return (
-    <AccordionItem value="모집 인원">
-      <AccordionTrigger titleIconName="group" summary={summary}>
+    <RecruitFormAccordion.Item value="모집 인원">
+      <RecruitFormAccordion.Trigger titleIconName="group" summary={summary}>
         모집 인원
-      </AccordionTrigger>
-      <AccordionContent>
+      </RecruitFormAccordion.Trigger>
+      <RecruitFormAccordion.Content>
         <Participants />
-      </AccordionContent>
-    </AccordionItem>
+      </RecruitFormAccordion.Content>
+    </RecruitFormAccordion.Item>
   );
 };
 
@@ -77,21 +72,21 @@ const EndDateLayer = () => {
   }) as RecruitFormValues['endDate'];
 
   return (
-    <AccordionItem value="모집 마감일 선택">
-      <AccordionTrigger titleIconName="calendar" summary={endDate}>
+    <RecruitFormAccordion.Item value="모집 마감일 선택">
+      <RecruitFormAccordion.Trigger titleIconName="calendar" summary={endDate}>
         모집 마감일 선택
-      </AccordionTrigger>
-      <AccordionContent>
+      </RecruitFormAccordion.Trigger>
+      <RecruitFormAccordion.Content>
         <EndDate />
-      </AccordionContent>
-    </AccordionItem>
+      </RecruitFormAccordion.Content>
+    </RecruitFormAccordion.Item>
   );
 };
 
 const SkillsLayer = () => {
   return (
-    <AccordionItem value="기술 스택">
-      <AccordionTrigger
+    <RecruitFormAccordion.Item value="기술 스택">
+      <RecruitFormAccordion.Trigger
         titleIconName="board"
         summary={
           <SelectedSkills
@@ -103,109 +98,10 @@ const SkillsLayer = () => {
         summaryContainerStyle={{ margin: 0 }}
       >
         기술 스택
-      </AccordionTrigger>
-      <AccordionContent>
+      </RecruitFormAccordion.Trigger>
+      <RecruitFormAccordion.Content>
         <Skills />
-      </AccordionContent>
-    </AccordionItem>
+      </RecruitFormAccordion.Content>
+    </RecruitFormAccordion.Item>
   );
 };
-
-//
-
-const accordionRootCss = css(flex('', '', 'column', 8));
-
-const accordionThemeCss = css({
-  backgroundColor: palettes.background.grey,
-  color: palettes.white,
-});
-
-interface AccordionItemProps {
-  value: string;
-  children: ReactNode;
-}
-
-const AccordionItem = (props: AccordionItemProps) => {
-  const { value, children } = props;
-  return (
-    <Accordion.Item
-      css={[accordionItemSelfCss, accordionThemeCss, recruitFormExpandCss]}
-      value={value}
-    >
-      {children}
-    </Accordion.Item>
-  );
-};
-
-const accordionItemSelfCss = css({
-  '&:focus-within': {
-    outline: `2px solid ${palettes.primary.default}`,
-  },
-});
-
-interface AccordionTriggerProps {
-  children: ReactNode;
-  titleIconName: IconNames;
-  summary?: ReactNode;
-  summaryContainerStyle?: CSSProperties;
-}
-
-const AccordionTrigger = (props: AccordionTriggerProps) => {
-  const { children, titleIconName, summary, summaryContainerStyle } = props;
-
-  return (
-    <Accordion.Trigger css={[accordionTriggerSelfCss, accordionThemeCss]}>
-      <div css={accordionTriggerRowCss}>
-        <div css={accordionTriggerTitleLayerCss}>
-          <Icon name={titleIconName} size={24} color={palettes.white} />
-          <div>{children}</div>
-        </div>
-        <Icon name="chevron.down" size={24} />
-      </div>
-      {summary && (
-        <div
-          css={[accordionTriggerRowCss, summaryCss]}
-          style={summaryContainerStyle}
-        >
-          {summary}
-        </div>
-      )}
-    </Accordion.Trigger>
-  );
-};
-
-const accordionTriggerSelfCss = css(
-  { width: '100%', padding: '6px 25px' },
-  fontCss.style.R16
-);
-
-const accordionTriggerRowCss = css(flex('center', 'space-between', 'row', 20));
-
-const accordionTriggerTitleLayerCss = css(
-  flex('center', 'flex-start', 'row', 8)
-);
-
-const summaryCss = css(
-  {
-    '[data-state="open"] > &': { display: 'none' },
-    '[data-state="closed"] > &': { display: '' },
-    marginTop: 6,
-  },
-  fontCss.style.R12
-);
-
-interface AccordionContentProps {
-  children: ReactNode;
-}
-const AccordionContent = (props: AccordionContentProps) => {
-  const { children } = props;
-  return (
-    <Accordion.Content>
-      <div css={accordionContentCss}>{children}</div>
-    </Accordion.Content>
-  );
-};
-
-const accordionContentCss = css({
-  padding: '12px 25px 120px',
-});
